requirements.txt
anthropic>=0.39.0
Pillow>=10.0.0
pyzbar>=0.1.9
openpyxl>=3.1.0
